adk-first-agent/requirements.txt
# Google ADK
google-adk==1.20.0

# Google GenAI
google-genai==1.55.0

# Database
psycopg2-binary==2.9.9

# Embeddings
sentence-transformers==3.0.1

# Environment
python-dotenv==1.2.1

# PyTorch (required for sentence-transformers)
torch==2.5.1
